App 1.3: Funcionalidad Boton

+++ App_1_3_Grupo_3_2024/Core/Inc/main.h
@@ -57,7 +57,7 @@ void Error_Handler(void);
 /* USER CODE BEGIN EFP */
 
 uint8_t secuencia1(uint16_t* vector_Pin_LEDs, uint8_t retardo_ms);
-uint8_t secuencia2(uint16_t* vector_Pin_LEDs, uint8_t retardo_ms);
+uint8_t secuencia2(uint16_t* vector_Pin_LEDs, uint16_t retardo_ms);
 uint8_t secuencia3(uint16_t* vector_Pin_LEDs);
 uint8_t secuencia4(uint16_t* vector_Pin_LEDs);
